refactor: changes from windowsx.zig refactoring

+++ Chapter 04 An Exercise in Text Output/03 SysMets3/SysMets3.zig
@@ -183,7 +183,7 @@ const Handler = struct {
         return 0;
     }
 
-    pub fn OnSize(self: *Handler, hwnd: HWND, _: u32, cx: i32, cy: i32) void {
+    pub fn OnSize(self: *Handler, hwnd: HWND, _: u32, cx: i16, cy: i16) void {
         self.client_width = cx;
         self.client_height = cy;
 
@@ -216,7 +216,7 @@ const Handler = struct {
         }
     }
 
-    pub fn OnVScroll(self: *Handler, hwnd: HWND, _: ?HWND, code: u32, _: i32) void {
+    pub fn OnVScroll(self: *Handler, hwnd: HWND, _: ?HWND, code: u16, _: i16) void {
         // Get all the vertical scroll bar information
         var si = SCROLLINFO{
             .cbSize = @sizeOf(SCROLLINFO),
@@ -259,7 +259,7 @@ const Handler = struct {
         }
     }
 
-    pub fn OnHScroll(self: *Handler, hwnd: HWND, _: ?HWND, code: u32, _: i32) void {
+    pub fn OnHScroll(self: *Handler, hwnd: HWND, _: ?HWND, code: u16, _: i16) void {
         // Get all the horizontal scroll bar information
         var si = SCROLLINFO{
             .cbSize = @sizeOf(SCROLLINFO),
