+++ Chapter 05 Basic Drawing/07 RandRect/RandRect.zig
@@ -165,7 +165,7 @@ const Handler = struct {
     cyClient: i32 = undefined,
     rnd: std.rand.DefaultPrng = RndGen.init(0),
 
-    pub fn OnSize(self: *Handler, _: HWND, _: u32, cx: i32, cy: i32) void {
+    pub fn OnSize(self: *Handler, _: HWND, _: u32, cx: i16, cy: i16) void {
         self.cxClient = cx;
         self.cyClient = cy;
     }
